Move known_first, known_step, known_last
