+++ src/ranges.jl
@@ -1,86 +1,4 @@
 
-_cartesian_index(i::Tuple{Vararg{Int}}) = CartesianIndex(i)
-_cartesian_index(::Any) = nothing
-
-"""
-    known_first(::Type{T}) -> Union{Int,Nothing}
-
-If `first` of an instance of type `T` is known at compile time, return it.
-Otherwise, return `nothing`.
-
-```julia
-julia> ArrayInterface.known_first(typeof(1:4))
-nothing
-
-julia> ArrayInterface.known_first(typeof(Base.OneTo(4)))
-1
-```
-"""
-known_first(x) = known_first(typeof(x))
-function known_first(::Type{T}) where {T}
-    if parent_type(T) <: T
-        return nothing
-    else
-        return known_first(parent_type(T))
-    end
-end
-known_first(::Type{Base.OneTo{T}}) where {T} = 1
-function known_first(::Type{T}) where {N,R,T<:CartesianIndices{N,R}}
-    _cartesian_index(ntuple(i -> known_first(R.parameters[i]), Val(N)))
-end
-
-"""
-    known_last(::Type{T}) -> Union{Int,Nothing}
-
-If `last` of an instance of type `T` is known at compile time, return it.
-Otherwise, return `nothing`.
-
-```julia
-julia> ArrayInterface.known_last(typeof(1:4))
-nothing
-
-julia> ArrayInterface.known_first(typeof(static(1):static(4)))
-4
-
-```
-"""
-known_last(x) = known_last(typeof(x))
-function known_last(::Type{T}) where {T}
-    if parent_type(T) <: T
-        return nothing
-    else
-        return known_last(parent_type(T))
-    end
-end
-function known_last(::Type{T}) where {N,R,T<:CartesianIndices{N,R}}
-    _cartesian_index(ntuple(i -> known_last(R.parameters[i]), Val(N)))
-end
-
-"""
-    known_step(::Type{T}) -> Union{Int,Nothing}
-
-If `step` of an instance of type `T` is known at compile time, return it.
-Otherwise, return `nothing`.
-
-```julia
-julia> ArrayInterface.known_step(typeof(1:2:8))
-nothing
-
-julia> ArrayInterface.known_step(typeof(1:4))
-1
-
-```
-"""
-known_step(x) = known_step(typeof(x))
-function known_step(::Type{T}) where {T}
-    if parent_type(T) <: T
-        return nothing
-    else
-        return known_step(parent_type(T))
-    end
-end
-known_step(::Type{<:AbstractUnitRange}) = 1
-
 """
     OptionallyStaticUnitRange(start, stop) <: AbstractUnitRange{Int}
